+++ pypeit/core/extract.py
@@ -2387,174 +2387,3 @@ def ech_objfind(image, ivar, slitmask, slit_left, slit_righ, order_vec, maskslit
 
     return sobjs_final, skymask[allmask]
 
-
-class ManualExtractionObj(pypeitpar.ParSet):
-    """
-    An object holding the arguments for how to perform the
-    manual extraction of a spectrum.
-
-    A list of these objects is generated in pypeit.py
-    to perform a set of user-defined extractions.
-
-    For an example of how to define a series of manual extractions in
-    the pypeit input file, see :ref:`pypeit_file`.
-
-    Args:
-        frame (:obj:`str`):
-            The name of the fits file for a manual extraction
-        spat (list): List of spatial, spectral positions to hand extract
-            e.g. "1243.3:1200," or "1243.3:1200,1345:1200'
-        spec (list): List of spatial, spectral positions to hand extract
-            e.g. "1243.3:1200," or "1243.3:1200,1345:1200'
-        det (list): List of detectors for hand extraction. This must be a list aligned with spec and spat lists, or a single integer
-             which will be used for all members of that list
-        fwhm (list): List of FWHM for hand extraction. This must be a list aligned with spec and spat lists, or a single number which will
-             be used for all members of that list'
-
-
-    """
-    @classmethod
-    def by_fitstbl_input(cls, frame: str, inp: str):
-        """Generate the object from an entry in the fitstbl
-
-        Args:
-            frame (str): filename
-            inp (str): det:spat:spec:fwhm
-                1:1181.8:3820.6:3.
-
-        Returns:
-            ManualExtractionObj or None:
-        """
-        if len(inp.strip()) == 0:
-            return None
-
-        # Generate a dict
-        idict = dict(frame=frame, spat=[], spec=[], det=[], fwhm=[])
-        m_es = inp.split(',')
-        for m_e in m_es:
-            parse = m_e.split(':')
-            idict['det'] += [int(parse[0])]
-            idict['spat'] += [float(parse[1])]
-            idict['spec'] += [float(parse[2])]
-            idict['fwhm'] += [float(parse[3])]
-
-        # Build me
-        return cls.from_dict(idict)
-
-
-    def __init__(self, frame=None, spat=None, spec=None, det=None, fwhm=None):
-
-        # Grab the parameter names and values from the function
-        # arguments
-        args, _, _, values = pypeitpar.inspect.getargvalues(pypeitpar.inspect.currentframe())
-        pars = pypeitpar.OrderedDict([(k,values[k]) for k in args[1:]])
-
-        # Initialize the other used specifications for this parameter
-        # set
-        dtypes = pypeitpar.OrderedDict.fromkeys(pars.keys())
-        descr = pypeitpar.OrderedDict.fromkeys(pars.keys())
-
-        # Fill out parameter specifications.  Only the values that are
-        # *not* None (i.e., the ones that are defined) need to be set
-        dtypes['frame'] = str
-        descr['frame'] = 'The name of the fits file for a manual extraction'
-
-        dtypes['spat'] = list # float
-        descr['spat'] = 'List of spatial positions to hand extract'
-
-        dtypes['spec'] = list # float
-        descr['spec'] = 'List of spectral positions to hand extract'
-
-        dtypes['det'] = list  # int
-        descr['det'] = 'List of detectors for hand extraction. This must be a list aligned with the spec_spat list.  Negative values indicate negative images.'
- 
-        dtypes['fwhm'] = list # float
-        descr['fwhm'] = 'List of FWHM (in pixels) for hand extraction. This must be a list aligned with spec_spat'
-
-        # Instantiate the object
-        super(ManualExtractionObj, self).__init__(list(pars.keys()),
-                                                  values=list(pars.values()),
-                                                  dtypes=list(dtypes.values()),
-                                                  descr=list(descr.values()))
-        self.validate()
-
-    @classmethod
-    def from_dict(cls, cfg):
-        k = np.array([*cfg.keys()])
-        parkeys = ['spat', 'spec','det','fwhm','frame']
-
-        badkeys = np.array([pk not in parkeys for pk in k])
-        if np.any(badkeys):
-            raise ValueError('{0} not recognized key(s) for ManualExtractionObj.'.format(
-                                k[badkeys]))
-
-        kwargs = {}
-        for pk in parkeys:
-            kwargs[pk] = cfg[pk] if pk in k else None
-        return cls(**kwargs)
-
-    def validate(self):
-        if self.data['spat'] is not None:
-            p1 = self.data['spat']
-            p2 = self.data['spec']
-            p3 = self.data['det']
-            p4 = self.data['fwhm']
-            if isinstance(p1, list):
-                if len(p1) != len(p2):
-                    raise ValueError("Each of these lists need the same length")
-                if len(p2) != len(p3):
-                    raise ValueError("Each of these lists need the same length")
-
-    def dict_for_objfind(self, neg=False):
-        """
-        Repackage into a dict for the extraction code
-
-        Args:
-            neg (bool, optional):
-                Negative image
-
-        Returns:
-            dict: To be passed into reduce.find_objects()
-
-        """
-        '''
-        if self.data['spat'] is None:
-            return None
-        if isinstance(self.data['det'], list):
-            spat_spec = self.data['spat_spec']
-            det = [int(obj) for obj in self.data['det']]
-            fwhm = [float(obj) for obj in self.data['fwhm']]
-        else:
-            spat_spec = [self.data['spat_spec']]
-            det = [self.data['det']]
-            fwhm = [self.data['fwhm']]
-        # Deal with spat_spec
-        spats, specs = [], []
-        for ispat_spec in spat_spec:
-            ps = ispat_spec.split(':')
-            spats.append(float(ps[0]))
-            specs.append(float(ps[1]))
-        # dict and return
-        '''
-        manual_dict =  dict(hand_extract_spec=self.data['spec'], 
-                    hand_extract_spat=self.data['spat'],
-                    hand_extract_det=self.data['det'], 
-                    hand_extract_fwhm=self.data['fwhm'])
-        #
-        dets = np.atleast_1d(manual_dict['hand_extract_det'])
-        # Grab the ones we want
-        gd_det = dets > 0
-        if not neg:
-            gd_det = np.invert(gd_det)
-        # Any?
-        if not np.any(gd_det):
-            return manual_dict
-        # Fill
-        manual_extract_dict = {}
-        for key in manual_dict.keys():
-            sgn = 1
-            if key == 'hand_extract_det':
-                sgn = -1
-            manual_extract_dict[key] = sgn*np.atleast_1d(manual_dict[key])[gd_det]
-        # Return
-        return manual_extract_dict
