+++ pypeit/par/__init__.py
@@ -1,4 +1,2 @@
 
 from pypeit.par.pypeitpar import PypeItPar
-#from pypeit.par.pypeitpar import ManualExtractionPar
-
