+++ pypeit/tests/test_pypeitpar.py
@@ -50,9 +50,6 @@ def test_sensfuncuvis():
 def test_telluric():
     pypeitpar.TelluricPar()
 
-#def test_manualextraction():
-#    pypeitpar.ManualExtractionPar()
-
 # TODO: Valid spectrographs are not longer read by pypeit.pypeitpar; it causes
 # a circular import.
 #def test_spectrographs():
